+++ doc/whats-new.md
@@ -64,6 +64,7 @@ A number of minor improvements:
  - there's now a legend under the overworld magnifier
  - the current dungeon tab highlights corresponding blockers area
  - dungeon doors/rooms give mouseover feedback, to make it easier to click the intended target
+ - there's an option to automatically infer some dungeon door marks (see [the documentation](use.md#main-dt) for Doors)
  - 'Show HotKeys' window remembers its size/location
  - added Z-Tracker logo next to the kitty
  - big icons checkbox in dungeons is persisted across sessions